Rename remote_ssh to remote_zfs
